requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==3.0.0
Django==2.2
django-allauth==0.53.0
django-cors-headers==3.2.0
django-oauth-toolkit==2.2.0
django-webpack-loader==1.8.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
ecdsa==0.18.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
jwcrypto==1.4.2
Markdown==3.1.1
numpy==1.18.0
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
social-auth-app-django==5.0.0
social-auth-core==4.4.0
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
